refactor: Move server browser fetch to util module
